Add some athutentications and apply the improved URLs view in CRUD views

+++ routes/web.php
@@ -13,20 +13,20 @@ use App\Http\Controllers\AnswerController;
 
 Route::get('/', [PageController::class, 'index'])->name('home');
 
-Route::get('questions', [QuestionController::class, 'index'])->name('questions.index');
+Route::get('foro', [QuestionController::class, 'index'])->name('questions.index');
 
-Route::get('questions/create', [QuestionController::class, 'create'])->name('questions.create'); //mostrar el diseno del formulario
-Route::post('questions', [QuestionController::class, 'store'])->name('questions.store'); //guardar la informacion del formulario en BD
+Route::get('foro/crear', [QuestionController::class, 'create'])->name('questions.create')->middleware('auth'); //mostrar el diseno del formulario
+Route::post('foro', [QuestionController::class, 'store'])->name('questions.store')->middleware('auth');; //guardar la informacion del formulario en BD
 
-Route::get('questions/{question}/edit', [QuestionController::class, 'edit'])->name('questions.edit'); //Muestra el formulario
-Route::put('questions/{question}', [QuestionController::class, 'update'])->name('questions.update'); //Actualiza la informacion en BD
-
-
-Route::get('questions/{question}', [QuestionController::class, 'show'])->name('questions.show');
-Route::delete('questions/{question}', [QuestionController::class, 'destroy'])->name('questions.destroy');
+Route::get('foro/{question:slug}/editar', [QuestionController::class, 'edit'])->name('questions.edit')->middleware('auth');; //Muestra el formulario
+Route::put('foro/{question:slug}', [QuestionController::class, 'update'])->name('questions.update')->middleware('auth', 'can:update,question'); //Actualiza la informacion en BD
 
 
-Route::post('/answers/{question}', [AnswerController::class, 'store'])->name('answers.store');
+Route::get('foro/{question:slug}', [QuestionController::class, 'show'])->name('questions.show');
+Route::delete('questions/{question:slug}', [QuestionController::class, 'destroy'])->name('questions.destroy')->middleware('auth', 'can:delete,question');
+
+
+Route::post('/answers/{question}', [AnswerController::class, 'store'])->name('answers.store')->middleware('auth');;
 
 
 Route::view('dashboard', 'dashboard')
